Move cache functions from eth to bounties
