+++ src/polyswarmd/utils/eth.py
@@ -11,7 +11,6 @@ from web3.exceptions import MismatchedABI
 from web3.module import Module
 from web3.utils.events import get_event_data
 
-from polyswarmd.app import cache
 from polyswarmd.websockets import messages
 
 logger = logging.getLogger(__name__)
@@ -97,12 +96,6 @@ def build_transaction(call, nonce):
     logger.debug('options: %s', options)
 
     return call.buildTransaction(options)
-
-
-@cache.memoize(1)
-def get_txpool():
-    from flask import g
-    return g.chain.w3.txpool.inspect
 
 
 def decode_all(raw_txs):
@@ -257,38 +250,3 @@ def events_from_transaction(txhash, chain):
                     continue
 
     return ret
-
-
-@cache.memoize(1)
-def bounty_fee(bounty_registry):
-    return bounty_registry.functions.bountyFee().call()
-
-
-@cache.memoize(1)
-def assertion_fee(bounty_registry):
-    return bounty_registry.functions.assertionFee().call()
-
-
-@cache.memoize(1)
-def bounty_amount_min(bounty_registry):
-    return bounty_registry.functions.BOUNTY_AMOUNT_MINIMUM().call()
-
-
-@cache.memoize(1)
-def assertion_bid_min(bounty_registry):
-    return bounty_registry.functions.ASSERTION_BID_ARTIFACT_MINIMUM().call()
-
-
-@cache.memoize(1)
-def assertion_bid_max(bounty_registry):
-    return bounty_registry.functions.ASSERTION_BID_ARTIFACT_MAXIMUM().call()
-
-
-@cache.memoize(1)
-def staking_total_max(arbiter_staking):
-    return arbiter_staking.functions.MAXIMUM_STAKE().call()
-
-
-@cache.memoize(1)
-def staking_total_min(arbiter_staking):
-    return arbiter_staking.functions.MINIMUM_STAKE().call()
